+++ TeamIntroduce/TeamIntroduce/Sources/Presnetaion/MemberDetail/MemberDetailView.swift
@@ -158,30 +158,6 @@ struct EditButton: View {
     }
 }
 
-
-struct CardStyle: ViewModifier {
-    func body(content: Content) -> some View {
-        content
-            .padding(15)
-            .frame(maxWidth: .infinity, alignment: .leading)
-            .background(
-                RoundedRectangle(cornerRadius: 16)
-                    .fill(.staticWhite)
-            )
-            .overlay(
-                RoundedRectangle(cornerRadius: 16)
-                    .stroke(.borderInactive, lineWidth: 0.5)
-            )
-            .shadow(color: .shadowDefault, radius: 3, x: 0, y: 1)
-    }
-}
-
-extension View {
-    public func cardStyle() -> some View {
-        modifier(CardStyle())
-    }
-}
-
 struct MemberDetailView_Previews: PreviewProvider {
     static var previews: some View {
         MemberDetailView()
